+++ composeApp/src/commonMain/kotlin/ru/llama/tool/domain/models/AIDialogChatDto.kt
@@ -1,7 +1,7 @@
 package ru.llama.tool.domain.models
 
 data class AIDialogChatDto(
-    val chatId: Int,
+    val chatId: Long,
     val chatName: String,
     val messages: List<Message>
 )
