+++ composeApp/src/commonMain/kotlin/ru/llama/tool/domain/use_cases/chat_property_interactor/ChatPropsInteractor.kt
@@ -5,9 +5,9 @@ import ru.llama.tool.domain.models.AiDialogProperties
 
 interface ChatPropsInteractor {
 
-    suspend fun getChatProperty(id: Int): Flow<AiDialogProperties>
+    suspend fun getChatProperty(id: Long): Flow<AiDialogProperties>
 
-    suspend fun deleteChatProperty(id: Int)
+    suspend fun deleteChatProperty(id: Long)
 
     suspend fun saveChatProperty(data: AiDialogProperties)
